fix: remove the FLUX identity that still broke Start
Start failed on a live Windows install for reasons the model swap had left
behind, all of them in the path the button actually takes.

The shipped default studio profile still named network volume ukh207b26r, which
no longer exists, so no Pod could be created. It now names the production volume.

The profile also carried model_preset flux2-klein-bf16, and the parser compared
that exact string, so any corrected profile would have been rejected anyway.

The configuration validator compared the model against Comfy-Org/Mage-Flow while
its error still read "must use FLUX.2 Klein 4B", which told anyone hitting it to
set precisely the value they already had.

The pinned generation contract described BF16 while the worker runs INT8
ConvRot, in the desktop contract, the JSON schema, the benchmark record contract
and its vectors. A contract that misreports precision cannot admit a benchmark
honestly, so all of them now say INT8.

+++ packages/runpod-client/src/benchmark-v2.ts
@@ -11,7 +11,7 @@ export interface GpuBenchmarkContractV2 {
   readonly modelRevision: string;
   readonly workerImageDigest: string;
   readonly templateId: string;
-  readonly precision: "bf16";
+  readonly precision: "int8";
   readonly width: 1280;
   readonly height: 720;
   readonly steps: 4;
@@ -114,7 +114,7 @@ export function parseGpuBenchmarkContractV2(value: unknown): GpuBenchmarkContrac
     typeof record.modelRevision !== "string" || record.modelRevision.length < 1 || record.modelRevision.length > 128 ||
     !isLowercaseRegistryImageDigestV1(record.workerImageDigest) ||
     typeof record.templateId !== "string" || record.templateId.length < 1 || record.templateId.length > 191 ||
-    record.precision !== "bf16" || record.width !== 1280 || record.height !== 720 ||
+    record.precision !== "int8" || record.width !== 1280 || record.height !== 720 ||
     record.steps !== 4 || record.guidanceMilli !== 1000 || record.jpegQuality !== 95 ||
     record.referenceMode !== "none" ||
     record.promptFixtureId !== "imageforge-gpu-benchmark-30-v1" ||
